first version fo app

+++ lib/data/models/current.dart
@@ -27,7 +27,7 @@ class Current {
     pressure = json['pressure'];
     humidity = json['humidity'];
     dewPoint = json['dew_point'];
-    uvi = json['uvi'];
+    uvi = json['uvi'].toDouble();
     clouds = json['clouds'];
     visibility = json['visibility'];
     windSpeed = json['wind_speed'];
